Update typescript and ts-node
The current typescript/ts-node combination is not working. Out of the
box runinng the tests fails with

```
Error: Debug Failure. False expression: Non-string value passed to `ts.resolveTypeReferenceDirective`, likely by a wrapping package working with an outdated `resolveTypeReferenceDirectives` signature. This is probably not a problem in TS itself.
```

Updating both packages resolves the issue.

Change-type: minor